scripts/get_path_list/requirements.txt
PyYAML >= 6.0
